# docs/swagger.yaml
basePath: /api
definitions:
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  secret.CreateRequest:
    properties:
      data:
        type: object
      secret_type:
        type: string
    required:
    - data
    - secret_type
    type: object
  secret.CreateResponse:
    properties:
      id:
        type: integer
    type: object
  secret.SecretResponse:
    properties:
      created_at:
        type: string
      data:
        type: object
      id:
        type: integer
      secret_type:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  secret.UpdateRequest:
    properties:
      data:
        type: object
    required:
    - data
    type: object
  user.LoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  user.RegisterRequest:
    properties:
      login:
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
    required:
    - login
    - password
    type: object
info:
  contact: {}
  description: HTTP API для управления зашифрованными секретами.
  title: GophKeeper API
  version: 0.0.1
paths:
  /api/secret:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/secret.SecretResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Список секретов пользователя
      tags:
      - secret
    post:
      consumes:
      - application/json
      parameters:
      - description: Новый секрет
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/secret.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/secret.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Создать секрет
      tags:
      - secret
  /api/secret/{id}:
    delete:
      parameters:
      - description: ID секрета
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Удалить секрет
      tags:
      - secret
    get:
      parameters:
      - description: ID секрета
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/secret.SecretResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить секрет по ID
      tags:
      - secret
    put:
      consumes:
      - application/json
      parameters:
      - description: ID секрета
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные секрета
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/secret.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Обновить секрет
      tags:
      - secret
  /api/user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Учетные данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Вход пользователя
      tags:
      - user
  /api/user/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Выход пользователя (очистка сессии)
      tags:
      - user
  /api/user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - user
swagger: "2.0"